fix: stats query pagination
