+++ x/zentp/keeper/query_mints_and_burns.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	"cosmossdk.io/collections"
-	"cosmossdk.io/math"
 	"github.com/Zenrock-Foundation/zrchain/v6/x/zentp/types"
 	"github.com/cosmos/cosmos-sdk/types/query"
 	"github.com/pkg/errors"
@@ -74,45 +73,4 @@ func (k Keeper) queryBridge(goCtx context.Context, store collections.Map[uint64,
 	}
 
 	return keys, pageRes, err
-}
-
-func (k Keeper) Stats(goCtx context.Context, req *types.QueryStatsRequest) (*types.QueryStatsResponse, error) {
-	if req == nil {
-		return nil, errors.New("request is nil")
-	}
-
-	mintKeys, _, err := k.queryBridge(goCtx, k.mintStore, nil, "", req.Denom, "", "", types.BridgeStatus_BRIDGE_STATUS_COMPLETED, 0, 0)
-	if err != nil {
-		return nil, err
-	}
-
-	burnKeys, _, err := k.queryBridge(goCtx, k.burnStore, nil, "", req.Denom, "", "", types.BridgeStatus_BRIDGE_STATUS_COMPLETED, 0, 0)
-	if err != nil {
-		return nil, err
-	}
-
-	totalMinted := math.ZeroInt()
-	mintsCount := uint64(0)
-	for _, mint := range mintKeys {
-		if req.Address == "" || mint.Creator == req.Address {
-			totalMinted = totalMinted.Add(math.NewIntFromUint64(mint.Amount))
-			mintsCount++
-		}
-	}
-
-	totalBurned := math.ZeroInt()
-	burnsCount := uint64(0)
-	for _, burn := range burnKeys {
-		if req.Address == "" || burn.RecipientAddress == req.Address {
-			totalBurned = totalBurned.Add(math.NewIntFromUint64(burn.Amount))
-			burnsCount++
-		}
-	}
-
-	return &types.QueryStatsResponse{
-		TotalMinted: totalMinted.Uint64(),
-		TotalBurned: totalBurned.Uint64(),
-		MintsCount:  mintsCount,
-		BurnsCount:  burnsCount,
-	}, nil
 }
